# Improve kmerge's perf