--- a/src/kmerge.rs
+++ b/src/kmerge.rs
@@ -2,7 +2,7 @@
 use size_hint;
 
 use std::cmp::Ordering;
-use std::collections::BinaryHeap;
+use std::mem::replace;
 
 macro_rules! clone_fields {
     ($name:ident, $base:expr, $($field:ident),+) => (
@@ -14,38 +14,32 @@ macro_rules! clone_fields {
     );
 }
 
-/// A non-empty sequence
+/// Head element and Tail iterator pair
 ///
 /// `PartialEq`, `Eq`, `PartialOrd` and `Ord` are implemented by comparing sequences based on
 /// first items (which are guaranteed to exist).
 ///
-/// The meanings of `PartialOrd` and `Ord` are reversed so as to turn the `BinaryHeap` used in
+/// The meanings of `PartialOrd` and `Ord` are reversed so as to turn the heap used in
 /// `KMerge` into a min-heap.
-struct NonEmpty<I>
+struct HeadTail<I>
     where I: Iterator
 {
     head: I::Item,
     tail: I,
 }
 
-impl<I> NonEmpty<I>
+impl<I> HeadTail<I>
     where I: Iterator
 {
-    /// Constructs a `NonEmpty` from an `Iterator`. Returns `None` if the `Iterator` is empty.
-    fn new(mut it: I) -> Option<NonEmpty<I>> {
+    /// Constructs a `HeadTail` from an `Iterator`. Returns `None` if the `Iterator` is empty.
+    fn new(mut it: I) -> Option<HeadTail<I>> {
         let head = it.next();
         head.map(|h| {
-            NonEmpty {
+            HeadTail {
                 head: h,
                 tail: it,
             }
         })
-    }
-
-    /// Returns the next item in the sequence. If more items remain, the remainder of the sequence
-    /// is returned as well, otherwise `None`.
-    fn next(self) -> (I::Item, Option<Self>) {
-        (self.head, NonEmpty::new(self.tail))
     }
 
     /// Hints at the size of the sequence, same as the `Iterator` method.
@@ -54,60 +48,91 @@ impl<I> NonEmpty<I>
     }
 }
 
-impl<I> Clone for NonEmpty<I>
+impl<I> Clone for HeadTail<I>
     where I: Iterator + Clone,
           I::Item: Clone
 {
     fn clone(&self) -> Self {
-        clone_fields!(NonEmpty, self, head, tail)
+        clone_fields!(HeadTail, self, head, tail)
     }
 }
 
-impl<I> PartialEq for NonEmpty<I>
+impl<I> PartialEq for HeadTail<I>
     where I: Iterator,
           I::Item: PartialEq
 {
-    fn eq(&self, other: &NonEmpty<I>) -> bool {
+    fn eq(&self, other: &HeadTail<I>) -> bool {
         self.head.eq(&other.head)
     }
 }
 
-impl<I> Eq for NonEmpty<I>
+impl<I> Eq for HeadTail<I>
     where I: Iterator,
           I::Item: Eq
 {}
 
-impl<I> PartialOrd for NonEmpty<I>
+impl<I> PartialOrd for HeadTail<I>
     where I: Iterator,
           I::Item: PartialOrd
 {
-    fn partial_cmp(&self, other: &NonEmpty<I>) -> Option<Ordering> {
+    fn partial_cmp(&self, other: &Self) -> Option<Ordering> {
         other.head.partial_cmp(&self.head)
     }
 
-    fn lt(&self, other: &NonEmpty<I>) -> bool {
+    fn lt(&self, other: &Self) -> bool {
         other.head.lt(&self.head)
     }
 
-    fn le(&self, other: &NonEmpty<I>) -> bool {
+    fn le(&self, other: &Self) -> bool {
         other.head.le(&self.head)
     }
 
-    fn gt(&self, other: &NonEmpty<I>) -> bool {
+    fn gt(&self, other: &Self) -> bool {
         other.head.gt(&self.head)
     }
 
-    fn ge(&self, other: &NonEmpty<I>) -> bool {
+    fn ge(&self, other: &Self) -> bool {
         other.head.ge(&self.head)
     }
 }
 
-impl<I> Ord for NonEmpty<I>
+impl<I> Ord for HeadTail<I>
     where I: Iterator,
           I::Item: Ord
 {
     fn cmp(&self, other: &Self) -> Ordering {
         other.head.cmp(&self.head)
+    }
+}
+
+/// Make `data` a heap (max-heap w.r.t T's Ord).
+fn heapify<T: Ord>(data: &mut [T]) {
+    for i in 0..data.len() / 2 {
+        sift_down(data, i);
+    }
+}
+
+/// Sift down element at `index` (`heap` is a max-heap wrt T's Ord).
+fn sift_down<T: Ord>(heap: &mut [T], index: usize) {
+    debug_assert!(index <= heap.len());
+    let mut pos = index;
+    let mut child = 2 * pos + 1;
+    // the `pos` conditional is to avoid a bounds check
+    while pos < heap.len() && child < heap.len() {
+        let right = child + 1;
+
+        // pick the bigger of the two children
+        if right < heap.len() && heap[child] < heap[right] {
+            child = right;
+        }
+
+        // sift down is done if we are already in order
+        if heap[pos] > heap[child] {
+            return;
+        }
+        heap.swap(pos, child);
+        pos = child;
+        child = 2 * pos + 1;
     }
 }
 
@@ -120,16 +145,20 @@ impl<I> Ord for NonEmpty<I>
 pub struct KMerge<I>
     where I: Iterator
 {
-    heap: BinaryHeap<NonEmpty<I>>,
+    heap: Vec<HeadTail<I>>,
 }
 
 /// Create a `KMerge` iterator.
-pub fn kmerge_new<I>(it: I) -> KMerge<<I::Item as IntoIterator>::IntoIter>
+pub fn kmerge_new<I>(iter: I) -> KMerge<<I::Item as IntoIterator>::IntoIter>
     where I: Iterator,
           I::Item: IntoIterator,
           <<I as Iterator>::Item as IntoIterator>::Item: Ord
 {
-    KMerge { heap: it.filter_map(|it| NonEmpty::new(it.into_iter())).collect() }
+    let (lower, _) = iter.size_hint();
+    let mut heap = Vec::with_capacity(lower);
+    heap.extend(iter.filter_map(|it| HeadTail::new(it.into_iter())));
+    heapify(&mut heap);
+    KMerge { heap: heap }
 }
 
 impl<I> Clone for KMerge<I>
@@ -148,11 +177,16 @@ impl<I> Iterator for KMerge<I>
     type Item = I::Item;
 
     fn next(&mut self) -> Option<Self::Item> {
-        self.heap.pop().map(|p| {
-            let (h, t) = p.next();
-            t.map(|t| self.heap.push(t));
-            h
-        })
+        if self.heap.is_empty() {
+            return None;
+        }
+        let result = if let Some(next_0) = self.heap[0].tail.next() {
+            replace(&mut self.heap[0].head, next_0)
+        } else {
+            self.heap.swap_remove(0).head
+        };
+        sift_down(&mut self.heap, 0);
+        Some(result)
     }
 
     fn size_hint(&self) -> (usize, Option<usize>) {