… src/adaptors.rs