@bluss 2016-02-23
cc @bsteinb 

Maybe I'm stealing your fun here, by implementing some improvements.

Thanks to @frankmcsherry for gracefully putting his code online so that I could read & copy it directly :smile:

The custom heap also makes us ready for a "kmerge_by" adaptor (user-supplied ordering closure).