@bluss 2016-02-23
Related to #98 (but there's still avenues to explore there).