@bsteinb 2016-02-23
Steal away! :+1: My fun is not going to stand in the way of progress.

Very impressive increase in performance for the two way merge. That seems to go well beyond what I achieved with the modified `BinaryHeap` from `std`. Could you say how much you gained by inlining-by-hand `HeadTail`s `next()` method?