@pczarn 2016-02-23
I had an idea for a struct similar to `HeadTail` which I called `PrePeeked`. Perhaps that's a better name.

With specializable associated types, we could avoid storing `head` for `slice::Iter` and access `iter.as_slice()[0]` instead.