@bluss 2016-02-23
Thanks for the input on this!